+++ packages/build-graphql/src/core/consts.ts
@@ -3,3 +3,5 @@ import { name } from '../../package.json';
 const exportSpecifier = 'typedefs';
 
 export const virtualModuleId = `${name}/${exportSpecifier}`;
+
+export const pluginName = 'unplugin-graphql';
