+++ packages/build-graphql/src/core/esbuild/esbuildSetup.ts
@@ -1,8 +1,8 @@
 import fs from 'node:fs/promises';
 import type { PluginBuild } from 'esbuild';
 import type { ILogger } from '../..';
+import { virtualModuleId } from '../consts';
 import { findFiles } from '../findFiles';
-import { virtualModuleId } from '../module';
 import type { ResolvedOptions } from '../types';
 
 export const esbuildSetup = (build: PluginBuild, options: ResolvedOptions, logger: ILogger) => {
